Update TODO
GUI tools would be nice in terms of scaling the number of signs, and updating them. Power saving is mostly handled by default with majority of potential savings to come from reducing display brightness or underclocking (not all boards will do this). Bedtime would be a nice feature to have.